+++ src/app/locales/en/components/archive-browser/-archive-state.js
@@ -1,0 +1,10 @@
+export default {
+  pending: 'Pending',
+  building: 'Building',
+  verifying: 'Verifying',
+  preserved: 'Preserved',
+  verification_failed: 'Verification&nbsp;failed',
+  failed: 'Failed',
+  purging: 'Purging',
+  unknown: 'Unknown',
+};